// frontend-angular/src/frontend-client/model/editOrderInitialData.ts
/**
 * Restaurant Import API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import {EditOrderInitialDataOrderDto} from './editOrderInitialDataOrderDto';


export interface EditOrderInitialData {
    order: EditOrderInitialDataOrderDto;
}